+++ services/GateWay/src/utils/code.hpp
@@ -1,0 +1,24 @@
+/******************************************************************************
+ *
+ * @file       code.hpp
+ * @brief      返回状态码集中定义
+ *
+ * @author     KBchulan
+ * @date       2025/12/02
+ * @history
+ ******************************************************************************/
+
+#ifndef CODE_HPP
+#define CODE_HPP
+
+#include <cstdint>
+
+namespace utils
+{
+
+constexpr std::int16_t SUCCESS = 0;
+constexpr std::int16_t URL_NOT_FOUND = 1;
+
+}  // namespace utils
+
+#endif  // CODE_HPP